+++ CHANGELOG.md
@@ -1,9 +1,18 @@
 # Change Log
 All notable changes to the "vscode-tomcat" extension will be documented in this file.
 
+# 0.11.3
+### Changed
+* Fixed [#226](https://github.com/adashen/vscode-tomcat/issues/226): Can't add Tomcat server on Linux
+* Fixed [#239](https://github.com/adashen/vscode-tomcat/issues/239): Can't select different JVM to run Tomcat instance
+
+# 0.11.2
+### Changed
+* Update the `vscode-extension-telemetry-wrapper` to `0.8.0`.
+
 # 0.11.1
 ### Fixed
-Account for case when looking at context XML [#218](https://github.com/adashen/vscode-tomcat/issues/#218)
+* Account for case when looking at context XML [#218](https://github.com/adashen/vscode-tomcat/issues/#218)
 
 # 0.11.0
 ### Added
@@ -109,7 +118,7 @@ Account for case when looking at context XML [#218](https://github.com/adashen/v
 * Force stop Tomcat when closing Visual Studio Code [#6](https://github.com/adashen/vscode-tomcat/issues/6)
 
 ## 0.2.1
-* Fix Overview page issue 
+* Fix Overview page issue
 
 ## 0.2.0
 ### Added